fix(glance): revert to stringData and update SOPS encrypted_regex
The previous fix changed stringData to data, but this caused Kubernetes
to reject the Secret because ENC[...] encrypted content is not valid base64.

SOPS encrypted_regex was set to ^(data)$ which only decrypts the 'data' field.
Updated to ^(data|stringData)$ to properly decrypt stringData fields.

This allows Flux to decrypt the SOPS-encrypted stringData content before
applying the Secret to the cluster.